+++ torch/_logging/__init__.py
@@ -9,6 +9,7 @@ from ._internal import (
     _init_logs,
     DEFAULT_LOGGING,
     getArtifactLogger,
+    LazyString,
     set_logs,
     warning_once,
 )
