+++ torch/utils/_traceback.py
@@ -1,4 +1,5 @@
 from types import TracebackType
+from typing import List, Optional
 import tempfile
 import traceback
 import contextlib
@@ -223,7 +224,7 @@ class CapturedTraceback:
         import torch._C._profiler
 
         # Directly populate tracebacks that already have cached summaries
-        rs = []
+        rs: List[Optional[List[str]]] = []
         delayed_idxs = []
         for i, tb in enumerate(tbs):
             if tb._summary is not None:
